feat: implement NFSFileSystem trait for OverlayFs
- Add NfsOverlay wrapper with Arc<Mutex<OverlayFs>> for thread safety
- Use tokio::task::spawn_blocking for all DB operations (rusqlite not Send)
- Implement all NFSFileSystem trait methods: lookup, getattr, setattr, read, write, create, mkdir, remove, rename, readdir, symlink, readlink
- Convert between internal types (Attrs, ItemType) and NFS types (fattr3, ftype3)
- Map errors to nfsstat3 error codes
- Add comprehensive async tests for all operations
- Fix db.rs usize cast issue in rename_by_path

+++ rust/src/db.rs
@@ -413,7 +413,7 @@ impl Database {
                 .execute(
                     "UPDATE inodes SET path = ? || SUBSTR(path, ?)
                      WHERE path LIKE ? || '%'",
-                    rusqlite::params![new_prefix, old_prefix.len() + 1, old_prefix],
+                    rusqlite::params![new_prefix, (old_prefix.len() + 1) as i64, old_prefix],
                 )
                 .wrap_err("failed to update descendant paths")?;
         }
